Getting better
OK, guys, I think I slowly I'm getting the hold of it - I have at least understood what's going on and I've set up v. small player brain (i.e. you control everything, but through an agent), and I think it's time to set it up a bit better to represent sort of a cheating bot (i.e. always getting the right Y axis for the ball) and freeze left to right movement (as current bots don't use it). It is possible to beat a cheating bot just by using forward/backwards movement (I did it plenty of times with a mouse) so I guess a good start is to train a bot to do that in order to see if we could beat the simplest bot using ML agents.

Additional remark:
Probably the bugs with resetting play after one of the players wins is that ball is always reseted to the attackers serve, regardles of what state the FSM is in, which causes  the machine to think that's another winning state and reset again. So I need sb to set up:
1. Either proper reset of ball to the proper player's position (e.g. Blue serve-blue serve-red serve-red serve-andsoOn) or
2. change the FSM so only one of the players serve regardless of ping pong rules.
I think this would be crucial when training the bots, what do you think?