Trained vs Non-trained catcher - WOW
Training with several reward heuristics lead to a good bot, that wins most of the time.
Reward heuristics were:
* 0.05 for successful aim (i.e. one that does not leat to a hit in the net).
* -0.3 for hit in the net
* 0.6 for a win
Also, since both bots were too slow, the speed had to be increased to 20 (from 8).
Unfortunately the AI has infered that there are some flaws with the normal catcher and aims only high balls on the right (so never hits the net) and waits for the other bot to make a mistake - hitting the net, which then leads to the non-trained bot chasing the ball before it had touched the table.