Latenight fixes with Sam
Removed the catcher script from player's racket and removed some code that broke after we removed that.

So we never really let player control the hits in the past, it was always a fucking catcher script hitting the ball.

Let's try to see if we can make it work decently before falling asleep...

Co-Authored-By: sjg206 <sjg206@users.noreply.github.com>